% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grandtab_detail_short}
\alias{grandtab_detail_short}
\title{Condensed raw GrandTab tables (most recent 10 years)}
\format{
An object of class \code{list} of length 11.
}
\source{
California Department of Fish and Wildlife GrandTab
}
\usage{
grandtab_detail_short
}
\description{
Same structure as \code{\link{grandtab_detail}} but containing only the
10 most recent run years of data. Used internally for faster rendering during
the CI update process.
}
\keyword{datasets}
