% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grandtab_raw.R
\name{get_grandtab_raw}
\alias{get_grandtab_raw}
\title{Retrieve raw GrandTab tables}
\usage{
get_grandtab_raw(table = NULL)
}
\arguments{
\item{table}{A table identifier. Accepts a \code{numeric} table number or 
vector of table numbers of length < 12. If \code{table} is \code{NULL}
(default), returns all 11 GrandTab tables as a named list of tibbles.}
}
\value{
A tibble (if \code{length(table)} = 1) or a named list of tibbles.
}
\description{
Returns unaltered GrandTab tables (almost) exactly as they appear in the 
online GrandTab PDF. Can be piped to \code{\link{get_column_notes}} to 
retrieve column footnotes as they appear in the GrandTab PDF.
}
\examples{
# Single table
get_grandtab_raw(2)

# Selected tables
get_grandtab_raw(2:4)

# All tables
get_grandtab_raw()

# Pipe to get_column_notes
# get_grandtab_raw(2) \%>\% get_column_notes("3/)
}
